Organize native compiler and plugin sources
