+++ emlx/Makefile
@@ -22,11 +22,11 @@ EMLX_LIB_DIR = $(PRIV_DIR)/mlx/lib
 # `Application.app_dir(:emlx, "priv/mlx/...")`.
 EMLX_INCLUDE_DIR = $(PRIV_DIR)/mlx/include
 EMLX_GENERATED_INCLUDE_DIR = $(PRIV_DIR)/include
-EMLX_COMPAT_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx_plugin_build_compat.hpp
-EMLX_PLUGIN_ABI_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx_plugin_abi.hpp
-EMLX_PLUGIN_TOOLCHAIN_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx_plugin_toolchain.hpp
+EMLX_COMPAT_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx/plugin/build_compat.hpp
+EMLX_PLUGIN_ABI_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx/plugin/abi.hpp
+EMLX_PLUGIN_TOOLCHAIN_HEADER = $(EMLX_GENERATED_INCLUDE_DIR)/emlx/plugin/toolchain.hpp
 EMLX_STAGING_DIR = $(BUILD_DIR)/compat-staging
-EMLX_STAGED_COMPAT_HEADER = $(EMLX_STAGING_DIR)/emlx_plugin_build_compat.hpp
+EMLX_STAGED_COMPAT_HEADER = $(EMLX_STAGING_DIR)/emlx/plugin/build_compat.hpp
 EMLX_DEPFILE_DIR = $(BUILD_DIR)/deps
 EMLX_BUILD_ID_TOOL = $(BUILD_DIR)/tools/emlx_build_identity
 EMLX_PACKAGED_BUILD_ID_TOOL = $(PRIV_DIR)/build_support/emlx_build_identity
@@ -40,7 +40,7 @@ MLX_INSTALL_DIR = $(MLX_DIR)
 
 $(info LIBMLX_ENABLE_DEBUG=$(LIBMLX_ENABLE_DEBUG))
 # Build flags
-CFLAGS = -fPIC -I$(call esc,$(ERTS_INCLUDE_DIR)) -I$(call esc,$(MLX_INCLUDE_DIR)) \
+CFLAGS = -fPIC -I c_src -I$(call esc,$(ERTS_INCLUDE_DIR)) -I$(call esc,$(MLX_INCLUDE_DIR)) \
          -I$(call esc,$(FINE_INCLUDE_DIR)) -Wall -std=c++20
 ifeq ($(LIBMLX_ENABLE_DEBUG),true)
 		CFLAGS += -g
@@ -72,8 +72,8 @@ endif
 MAKE_JOBS ?= $(MAKE_DEFAULT_JOBS)
 
 # Source files
-SOURCES = c_src/emlx_nif.cpp c_src/emlx_fast.cpp c_src/emlx_compiler.cpp c_src/emlx_plugin_registry.cpp c_src/emlx_native_image.cpp c_src/emlx_nif_lifecycle.cpp c_src/emlx_sha256.cpp
-HEADERS = c_src/nx_nif_utils.hpp c_src/emlx_worker.hpp c_src/emlx_async.hpp c_src/emlx_nif_shared.hpp c_src/emlx_plugin_abi.hpp c_src/emlx_plugin_toolchain.hpp c_src/emlx_plugin_registry.hpp c_src/emlx_native_image.hpp c_src/emlx_nif_lifecycle.hpp c_src/emlx_sha256.hpp c_src/emlx_depfile.hpp c_src/emlx_compiler.hpp c_src/emlx_runtime_call_bridge.hpp
+SOURCES = c_src/emlx_nif.cpp c_src/emlx_fast.cpp c_src/emlx/compiler.cpp c_src/emlx/plugin/registry.cpp c_src/emlx_native_image.cpp c_src/emlx_nif_lifecycle.cpp c_src/emlx_sha256.cpp
+HEADERS = c_src/nx_nif_utils.hpp c_src/emlx_worker.hpp c_src/emlx_async.hpp c_src/emlx_nif_shared.hpp c_src/emlx/plugin/abi.hpp c_src/emlx/plugin/toolchain.hpp c_src/emlx/plugin/registry.hpp c_src/emlx_native_image.hpp c_src/emlx_nif_lifecycle.hpp c_src/emlx_sha256.hpp c_src/emlx_depfile.hpp c_src/emlx/compiler.hpp c_src/emlx_runtime_call_bridge.hpp
 OBJECTS = $(patsubst c_src/%.cpp,$(call esc,$(BUILD_DIR))/%.o,$(SOURCES))
 DEPFILES = $(patsubst c_src/%.cpp,$(call esc,$(EMLX_DEPFILE_DIR))/%.d,$(SOURCES))
 HOST_DEPFILE_ARGS = $(foreach source,$(SOURCES),"$(source)" "$(EMLX_DEPFILE_DIR)/$(patsubst c_src/%.cpp,%.d,$(source))")
@@ -85,7 +85,7 @@ all: $(call esc,$(MLX_SO)) $(call esc,$(EMLX_COMPAT_HEADER)) $(call esc,$(EMLX_P
 $(call esc,$(PRIV_DIR)):
 	@ mkdir -p "$(PRIV_DIR)"
 
-$(call esc,$(EMLX_BUILD_ID_TOOL)): c_src/tools/emlx_build_identity.cpp c_src/emlx_sha256.cpp c_src/emlx_sha256.hpp c_src/emlx_depfile.cpp c_src/emlx_depfile.hpp c_src/emlx_plugin_toolchain.hpp
+$(call esc,$(EMLX_BUILD_ID_TOOL)): c_src/tools/emlx_build_identity.cpp c_src/emlx_sha256.cpp c_src/emlx_sha256.hpp c_src/emlx_depfile.cpp c_src/emlx_depfile.hpp c_src/emlx/plugin/toolchain.hpp
 	@ mkdir -p "$(BUILD_DIR)/tools"
 	$(CXX) -std=c++20 -O2 -I c_src c_src/tools/emlx_build_identity.cpp c_src/emlx_sha256.cpp c_src/emlx_depfile.cpp -o "$@"
 
@@ -97,16 +97,16 @@ $(call esc,$(EMLX_PLUGIN_TOOL_WRAPPER)): c_src/tools/emlx_plugin_tool_wrapper.cp
 	@ mkdir -p "$(PRIV_DIR)/build_support"
 	$(CXX) -std=c++20 -O2 c_src/tools/emlx_plugin_tool_wrapper.cpp -o "$@"
 
-$(call esc,$(EMLX_PLUGIN_ABI_HEADER)): c_src/emlx_plugin_abi.hpp
-	@ mkdir -p "$(EMLX_GENERATED_INCLUDE_DIR)"
+$(call esc,$(EMLX_PLUGIN_ABI_HEADER)): c_src/emlx/plugin/abi.hpp
+	@ mkdir -p "$(dir $(EMLX_PLUGIN_ABI_HEADER))"
 	@ cp "$<" "$@"
 
-$(call esc,$(EMLX_PLUGIN_TOOLCHAIN_HEADER)): c_src/emlx_plugin_toolchain.hpp
-	@ mkdir -p "$(EMLX_GENERATED_INCLUDE_DIR)"
+$(call esc,$(EMLX_PLUGIN_TOOLCHAIN_HEADER)): c_src/emlx/plugin/toolchain.hpp
+	@ mkdir -p "$(dir $(EMLX_PLUGIN_TOOLCHAIN_HEADER))"
 	@ cp "$<" "$@"
 
 $(call esc,$(EMLX_STAGED_COMPAT_HEADER)): $(call esc,$(MLX_SO)) $(call esc,$(EMLX_BUILD_ID_TOOL)) FORCE
-	@ mkdir -p "$(EMLX_STAGING_DIR)"
+	@ mkdir -p "$(dir $(EMLX_STAGED_COMPAT_HEADER))"
 	"$(EMLX_BUILD_ID_TOOL)" "$(MLX_SO)" "$(MLX_INCLUDE_DIR)" "$(MLX_VERSION)" "$(MLX_VARIANT)" "$(EMLX_TARGET_TRIPLE)" "$(EMLX_STAGED_COMPAT_HEADER)"
 
 $(call esc,$(EMLX_DEPFILE_DIR))/%.d: c_src/%.cpp $(HEADERS) $(call esc,$(EMLX_STAGED_COMPAT_HEADER))
@@ -118,6 +118,7 @@ $(call esc,$(BUILD_DIR))/%.o: $(call esc,$(EMLX_DEPFILE_DIR))/%.d
 	@ test -f "$@"
 
 $(call esc,$(EMLX_COMPAT_HEADER)): $(OBJECTS) $(DEPFILES) $(call esc,$(EMLX_BUILD_ID_TOOL))
+	@ mkdir -p "$(dir $(EMLX_COMPAT_HEADER))"
 	@ rm -f "$(EMLX_COMPAT_HEADER)"
 	"$(EMLX_BUILD_ID_TOOL)" verify-host-deps "$(MLX_INCLUDE_DIR)" "$(EMLX_STAGED_COMPAT_HEADER)" "$(EMLX_COMPAT_HEADER)" $(HOST_DEPFILE_ARGS)
