+++ emlx/c_src/emlx/compiler.hpp
@@ -1,14 +1,14 @@
 #pragma once
 
-// emlx_compiler.hpp — emlx::native namespace: Op enum, Expr program struct,
+// emlx/compiler.hpp — emlx::native namespace: Op enum, Expr program struct,
 // and NIF implementation declarations for compile_program / eval_program /
 // native_expr_opcode_table.
 //
-// The three *_impl functions are implemented in emlx_compiler.cpp and called
+// The three *_impl functions are implemented in emlx/compiler.cpp and called
 // from thin NIF wrappers in emlx_nif.cpp.
 
 #include "emlx_nif_shared.hpp"
-#include "emlx_plugin_registry.hpp"
+#include "emlx/plugin/registry.hpp"
 #include <variant>
 
 namespace emlx {
@@ -28,7 +28,7 @@ namespace native {
 // loop-carry slot, respectively. `Carry` only ever appears inside a
 // `SubProgram` (a `:while` instruction's `cond`/`body`) — never in the
 // top-level program's own instruction list, and vice versa for `Input`
-// (see EMLXWhile's doc comment in emlx_compiler.cpp).
+// (see EMLXWhile's doc comment in emlx/compiler.cpp).
 enum class RefKind { Input, Capture, Const, Result, Carry };
 
 struct Ref {
@@ -41,7 +41,7 @@ struct Ref {
 // mode strings, sent as atoms so no int<->meaning lookup table needs to be
 // kept in sync with Elixir (see string2dtype/dtype_map in
 // emlx_nif_shared.hpp). The implicit int64_t conversion below keeps every
-// existing `attrs[i]`-as-int64_t use site in emlx_compiler.cpp's op registry
+// existing `attrs[i]`-as-int64_t use site in emlx/compiler.cpp's op registry
 // compiling unchanged.
 class Attr {
 public:
@@ -75,7 +75,7 @@ struct Instruction;
 // A `:while` instruction's condition or body, lowered as its own
 // self-contained flat instruction list (not inlined into the parent
 // program) — see EMLX.Native.Expr's `:while` moduledoc section and
-// EMLXWhile (emlx_compiler.cpp). `instructions`' own `Ref::Result` entries
+// EMLXWhile (emlx/compiler.cpp). `instructions`' own `Ref::Result` entries
 // are local to this sub-program (a fresh flat accumulator per interpretation
 // — see `interpret_instructions`), distinct from the parent program's own
 // `{:result, i}` numbering. `Ref::Carry` entries resolve against whatever
@@ -139,7 +139,7 @@ struct Expr {
 };
 
 // compile_program is defined via FINE_ASYNC_NIF(compile_program) in
-// emlx_compiler.cpp (declares `compile_program`/`compile_program_async`
+// emlx/compiler.cpp (declares `compile_program`/`compile_program_async`
 // here); eval_program is a plain hand-written NIF, called from a thin
 // wrapper in emlx_nif.cpp.
 ERL_NIF_TERM compile_program(ErlNifEnv *env, int argc,
