+++ emlx/test/emlx/native_identity_test.exs
@@ -215,13 +215,13 @@ defmodule EMLX.NativeIdentityTest do
     policy = Path.join(axon_root, "c_src/policy.txt")
     makefile = Path.join(axon_root, "Makefile")
     emlx_public = Path.join(axon_root, "_build/#{build_env}/lib/emlx/priv/include")
-    abi = Path.join(emlx_public, "emlx_plugin_abi.hpp")
-    packaged_toolchain = Path.join(emlx_public, "emlx_plugin_toolchain.hpp")
+    abi = Path.join(emlx_public, "emlx/plugin/abi.hpp")
+    packaged_toolchain = Path.join(emlx_public, "emlx/plugin/toolchain.hpp")
     wrapper = Path.join(emlx_root, "c_src/tools/wrapper.cpp")
-    source_toolchain = Path.join(emlx_root, "c_src/emlx_plugin_toolchain.hpp")
+    source_toolchain = Path.join(emlx_root, "c_src/emlx/plugin/toolchain.hpp")
     mlx_header = Path.join(mlx_root, "mlx/array.h")
     generated = Path.join(root, "generated")
-    compat = Path.join(emlx_public, "emlx_plugin_build_compat.hpp")
+    compat = Path.join(emlx_public, "emlx/plugin/build_compat.hpp")
     actual_mlx = Path.join(generated, "emlx_qwen3_mlx_headers_build_id.hpp")
     scan_header = Path.join(generated, "scan/emlx_qwen3_plugin_build_id.hpp")
     output_header = Path.join(generated, "final/emlx_qwen3_plugin_build_id.hpp")
