+++ emlx/test/emlx/native_lifecycle_test.exs
@@ -195,7 +195,7 @@ defmodule EMLX.NativeLifecycleTest do
       Path.join(root, "test/support/native_lifecycle_fixture.cpp"),
       Path.join(root, "c_src/emlx_nif_lifecycle.cpp"),
       Path.join(root, "c_src/emlx_native_image.cpp"),
-      Path.join(root, "c_src/emlx_plugin_registry.cpp"),
+      Path.join(root, "c_src/emlx/plugin/registry.cpp"),
       Path.join(root, "c_src/emlx_sha256.cpp"),
       "-I#{Path.join(root, "deps/fine/c_include")}",
       "-L#{mlx_lib}",
