+++ emlx/test/support/plugin_abi_layout.cpp
@@ -1,4 +1,4 @@
-#include "emlx_plugin_abi.hpp"
+#include "emlx/plugin/abi.hpp"
 
 #include <cstdint>
 #include <iostream>
